Split XDF EF assets into compat pbo
