+++ xdf_ef/CfgPatches.hpp
@@ -1,0 +1,18 @@
+class CfgPatches 
+{
+	class xdf_ef
+	{
+		requiredAddons[] = 
+		{
+			"xdf",
+			"ef"
+		};
+		version = "1.0";
+		author="Radium";
+		authorUrl = "https://steamcommunity.com/id/radiumio/";
+		units[] = {};
+		weapons[] = {};
+		
+		skipWhenMissingDependencies = 1;
+	};
+};
