+++ xdf_ef/config.cpp
@@ -1,0 +1,3 @@
+#include "CfgPatches.hpp"
+#include "CfgVehicles.hpp"
+#include "CfgWeapons.hpp"
